--- openmetadata-airflow-apis/openmetadata_managed_apis/api/api_setup.py
@@ -1,9 +1,0 @@
-from openmetadata_managed_apis.api.app import blueprint
-
-
-# Alas
-# import endpoints
-from openmetadata_managed_apis.api.routes import *
-
-# import error handlers
-from openmetadata_managed_apis.api import error_handlers
