+++ openmetadata-airflow-apis/openmetadata_managed_apis/api/routes/__init__.py
@@ -1,5 +1,5 @@
-from os.path import dirname, basename, isfile, join
 import glob
+from os.path import basename, dirname, isfile, join
 
 modules = glob.glob(join(dirname(__file__), "*.py"))
 __all__ = [
