+++ openmetadata-airflow-apis/openmetadata_managed_apis/operations/delete.py
@@ -17,7 +17,10 @@ from pathlib import Path
 from airflow import settings
 from airflow.models import DagModel, DagRun
 from flask import Response
-from openmetadata_managed_apis.api.config import AIRFLOW_DAGS_FOLDER, DAG_GENERATED_CONFIGS
+from openmetadata_managed_apis.api.config import (
+    AIRFLOW_DAGS_FOLDER,
+    DAG_GENERATED_CONFIGS,
+)
 from openmetadata_managed_apis.api.response import ApiResponse
